refactor(mobile): extract scheme-handler bridge into TriliumSchemeHandler.swift
Move TriliumAssetHandler, TriliumSchemeBridge, and the local API prefix list
out of ViewController.swift into a dedicated file; the view controller keeps
only the webView(with:configuration:) hook that isa-swizzles Capacitor's
registered handler. Registers the new file in the Xcode project.
